pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-simple"
version = "0.1.0"
description = "Simple RAG system - Portfolio project"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "chromadb>=0.6.0",
    "sentence-transformers>=3.3.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
